requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.8.0
asttokens==3.0.0
astunparse==1.6.3
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comm==0.2.2
cycler==0.12.1
debugpy==1.8.12
decorator==5.1.1
executing==2.2.0
fastapi==0.115.8
filelock==3.16.1
flatbuffers==25.2.10
fonttools==4.55.3
fsspec==2024.12.0
gast==0.6.0
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.70.0
h11==0.14.0
h5py==3.12.1
httpcore==1.0.7
httpx==0.28.1
idna==3.10
ipykernel==6.29.5
ipython==8.32.0
jedi==0.19.2
Jinja2==3.1.5
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==3.6.0
kiwisolver==1.4.8
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.4.1
namex==0.0.8
nest-asyncio==1.6.0
numpy==1.26.4
opt_einsum==3.4.0
optree==0.14.0
packaging==24.2
parso==0.8.4
pillow==10.4.0
platformdirs==2.6.2
prompt_toolkit==3.0.50
protobuf==4.25.6
psutil==6.1.1
pure_eval==0.2.3
py4j==0.10.9.7
pydantic==2.10.5
pydantic_core==2.27.2
Pygments==2.19.1
PyMySQL==1.1.1
pyparsing==3.2.1
pyspark==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.1
requests==2.32.3
rich==13.9.4
session_info==1.0.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
stack-data==0.6.3
starlette==0.45.3
stdlib-list==0.11.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.1
tensorflow-intel==2.17.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.5.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
uvicorn==0.34.0
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.17.2
